simulate.x: observe and foot subcommand work
observe: Once again, changed the way modules are specified in 'observe.conf'
files, to make interfacing with the web UI easier. The old way still works
but is deprecated. The modules should now be specified in the following
manner:

	module.module_name = 1
	module.module_name.param1 = val1
	module.module_name.param2 = val2
	etc...

The first line enables or disables the module 'module_name', while the
others specify any configuration settings for the module. When module class
is actually initiallized, the Config object passed to it will have the
module.module_name prefixes stripped from the key names.

foot: Added a 'foot' subcommand, which unifies all footprint generation
options. The type of footprint is specified in 'type' keyword of foot.conf
file. Added a 'type=rect' type, for selection of a rectangle in a given
coordinate system. NOTE: the functionality of 'footprint' subcommand has not
yet been implemented here (e.g., as 'type=sdssrun' or something similar).