simulate.x: Marsaglia's Multiply-With-Carry PRNG for GPU
The RNG as implemented allows for 64k independent streams of random numbers
(typically, accessed as one per thread). It's seeded from CPU RNG on first
use (and that is how it depends on user-defined seed). Note that if there
are more than 64k threads that need random numbers, they either have to
cooperate to avoid simultaneous drawing from the same stream, or something
else needs to be done.