KinTMIII module for cude working. See further comments.
KinTMIII module was converted to cuda. Trivariate gaussian rng generator
was coded and tested.

Three KinTMIII modules are included in code:
  KinTMIII_OLD - old version
  KinTMIII - transitional version
  otable version

First, the KinTMIII_OLD was compared to KinTMIII (both using gsl rng), and produced the same results.
Then rng for KinTMIII was swaped for Marsaglia's, and compared
with otable version, producing the same results (in first 128 rows).

Finally, otable version was run on cuda device, producing the same
results.

This commit has otable module set for execution, and KinTMIII is set to
Marsaglia's rng.

otable version of KinTMIII in this commit is not optimized, and needs
cleanup, but it is easy to compare it with previous versions.

Also, to avoid CUDA 'parameter overflow', module parameters were passed
to kernel with extremely limited precision(2 decimals, and only
integers for halo). Constant memory should be used instead.

Miraculously, the component==0 bug disappeared. But FeH module still has
if (component==0) FeH=0 code, for testing (should be removed).