Reorganize CUDA sources and fix dependency tracking
Reorganized all CUDA sources to be living in .h files, but be compiled in a
single .cu file. The former helps with correct syntax highlighting in
KDevelop, while the later is necessary because __constant__ variables have
static scope and if declared in multiple header files, that are included in
multiple .cu files, may create multiple symbols with the same name. This (at
runtime!) confuses cudaMemcpyToSymbol() and may produce nasty effects.