os_photometry: rework the meaning of output flags
Now the flags designate whether a given band has been computed from an
extrapolated color or not. For example, if the bands are ugrizy, r is the
bootstrap band, and gr color was looked up from an extrapolated region, then
the bands u and g are uncertain; the flags will be b110000.