Convert XYZ to Galactic (heliocentric!) cartesian coords
The origin of XYZ coordinates used to be the Galactic center. Since the
exact location of Galactic center is unknown, this introduced an additional
global parameter that had to be tracked through the code (Rg()). This also
introduced Rg into the basic machinery of density->stars Monte Carlo code,
where there's no need for it.

With this change, the code becomes independent of the location of the
Galactic center. The XYZ coordinates are simply the cartesian representation of
the usual Galactic coordinate system on the sky (plus distance). Therefore,
the Galactic center is in the +X direction, +Z looks towards the NGP, and +Y
is in the direction of Galactic rotation.

Density models and postprocessing modules that require the knowledge of
where the GC is should specify it as a parameter in their configuration.